requirements.txt
streamlit
pandas
numpy
pydeck
google-generativeai
